requirements.txt
fastapi>=0.115.0,<0.120.0
uvicorn[standard]>=0.32.0,<0.33.0
python-multipart>=0.0.9
# torch>=2.5.0,<2.6.0
# torchaudio>=2.5.0,<2.6.0

transformers>=4.45.0,<5.0.0
accelerate>=0.34.0
nemo-toolkit[asr]==2.5.1
megatron-core>=0.9.0

omegaconf>=2.3.0,<3.0.0
hydra-core>=1.3.0,<2.0.0
einops>=0.8.0
sentencepiece>=0.1.99
protobuf>=3.20.0

librosa>=0.10.2
soundfile>=0.12.1
pydub>=0.25.1

numpy>=1.26.0,<2.0.0
scipy>=1.13.0

sqlalchemy>=2.0.44,<3.0.0
alembic>=1.14.0

huggingface-hub>=0.26.0
safetensors>=0.4.5

python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
requests>=2.31.0

psutil>=6.1.0
tqdm>=4.67.0
cuda-python>=12.3
onnxruntime
pynvml>=11.5.0